Refactor : Renamed WISH to WISHLIST and moved reducer function to seperate file
