+++ src/Context/ActionTypes.js
@@ -1,10 +1,10 @@
-export const ADD_TO_WISH = "ADD_TO_WISH";
-export const MOVE_TO_CART_FROM_WISH = "MOVE_TO_CART_FROM_WISH";
-export const MOVE_TO_WISH_FROM_CART = "MOVE_TO_WISH_FROM_CART";
-export const REMOVE_FROM_WISH = "REMOVE_FROM_WISH";
+export const ADD_TO_WISHLIST = "ADD_TO_WISHLIST";
+export const MOVE_TO_CART_FROM_WISHLIST = "MOVE_TO_CART_FROM_WISHLIST";
+export const MOVE_TO_WISHLIST_FROM_CART = "MOVE_TO_WISHLIST_FROM_CART";
+export const REMOVE_FROM_WISHLIST = "REMOVE_FROM_WISHLIST";
 export const REMOVE_FROM_CART = "REMOVE_FROM_CART";
 export const ADD_TO_CART = "ADD_TO_CART";
-export const INCREASE_ITEM_IN_WISH = "INCREASE_ITEM_IN_WISH";
-export const DECREASE_ITEM_IN_WISH = "DECREASE_ITEM_IN_WISH";
+export const INCREASE_ITEM_IN_WISHLIST = "INCREASE_ITEM_IN_WISHLIST";
+export const DECREASE_ITEM_IN_WISHLIST = "DECREASE_ITEM_IN_WISHLIST";
 export const INCREASE_ITEM_IN_CART = "INCREASE_ITEM_IN_CART";
 export const DECREASE_ITEM_IN_CART = "DECREASE_ITEM_IN_CART";
